cronjob-copy-langfuse-prompts/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
langfuse==2.54.1
packaging==24.2
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing-extensions==4.13.2
urllib3==2.2.3
wrapt==1.17.2
